requirements.txt
pip>=21.2.4
tracardi-plugin-sdk
tracardi~=0.6.16
wheel
twine

setuptools~=57.0.0
pydantic~=1.8.2
boto3~=1.18.52
asyncio~=3.4.3
python-dotenv~=0.19.0
aiobotocore-1.4.2
aioitertools-0.8.0
botocore-1.20.106
wrapt-1.13.1

aiobotocore~=1.4.2